+++ src/main/clojure/clara/macros.clj
@@ -43,33 +43,25 @@
    (coll? source) (seq source)
    :else (throw (IllegalArgumentException. "Unknown source value type passed to defsession"))))
 
+(defn defrule!
+  [name production]
+  (add-production name production)
+  `(def ~name
+     ~production))
+
 (defmacro defrule
   [name & body]
-  (let [doc (if (string? (first body)) (first body) nil)
-        body (if doc (rest body) body)
-        properties (if (map? (first body)) (first body) nil)
-        definition (if properties (rest body) body)
-        {:keys [lhs rhs]} (dsl/split-lhs-rhs definition)
+  (defrule! name (dsl/build-rule name body)))
 
-        production (cond-> (dsl/parse-rule* lhs rhs properties {})
-                           name (assoc :name (str (clojure.core/name (com/cljs-ns)) "/" (clojure.core/name name)))
-                           doc (assoc :doc doc))]
-    (add-production name production)
-    `(def ~name
-       ~production)))
+(defn defquery!
+  [name query]
+  (add-production name query)
+  `(def ~name
+     ~query))
 
 (defmacro defquery
   [name & body]
-  (let [doc (if (string? (first body)) (first body) nil)
-        binding (if doc (second body) (first body))
-        definition (if doc (drop 2 body) (rest body) )
-
-        query (cond-> (dsl/parse-query* binding definition {})
-                      name (assoc :name (str (clojure.core/name (com/cljs-ns)) "/" (clojure.core/name name)))
-                      doc (assoc :doc doc))]
-    (add-production name query)
-    `(def ~name
-       ~query)))
+  (defquery! name (dsl/build-query name body)))
 
 (sc/defn gen-beta-network :- [sc/Any] ; Returns a sequence of compiled nodes.
   "Generates the beta network from the beta tree. "